Rolling back to an earlier version of the Tabular function
There was a newer version of the to_tabular funcion that used jsonbuffer. That worked for csvs but not for json. Reverting to the earlier function seems to work for everything. Here is the newer code that was reverted https://github.com/18F/ReVAL/blob/master/data_ingest/utils.py#L44C1-L87C30.
This commit makes the tests pass in the USDA-FNS app.